db_writer/requirements.txt
psycopg2-binary==2.9.9
confluent-kafka==2.3.0